Fix two sync defects: late FFT windows, and a float32 prefix-sum collapse
Both found while measuring the Schmidl-Cox timing metric, and unrelated
to each other.

1. Place every OFDM FFT window 2 samples early (Ofdm(timing_advance=)).

A cyclic prefix makes the window position free in the EARLY direction
only: starting inside the CP is a true cyclic shift, costing a phase
ramp that the channel estimate absorbs exactly. One sample LATE is not a
cyclic shift -- the window runs into the next symbol, and that leakage
depends on the neighbouring symbol's data, so no per-subcarrier H[k] can
represent it.

The metric's flat top is only 2 samples wide (the preamble has no CP, so
there is no broad plateau): M(0)=1.0000 vs M(1)=0.9995 noiseless. An echo
biases that 0.05% contest late, and MORE SNR makes it worse -- a cleaner
metric resolves a peak the echo has genuinely shifted, so at a=0.3 it
lands late 61% of the time at 20 dB and 100% at 40 dB. Static-channel
mean EVM, natural sync: a=0.3/40 dB goes 0.149 -> 0.014, a=0.2/25 dB
0.125 -> 0.073, and a=0.0 is unchanged at 0.071 (early really is free).
Uncoded 16QAM at a >= 0.2 goes from losing frames outright to decoding
bit-exactly.

Applied at `pos`, not to start_index, so the CFO estimator keeps
correlating at the preamble position sync detected. Training, header,
DMRS and payload all walk forward from that one `pos`, so they move
together. Default 2, auto-clamped for small/zero cp_len.

2. Accumulate the vectorized sync path's prefix sums in float64.

The two shipped paths of SchmidlCoxSync disagreed. The vectorized path
recovers each windowed sum by differencing two whole-buffer prefix sums,
so the accumulator's ulp -- not the window's magnitude -- sets the
smallest window that survives. In float32 on a 44,672-sample frame the
energy total reaches ~137 (ulp ~1.6e-5) while a sample past the frame is
~3.5e-7, ~46,000x below half-ulp: those samples stopped accumulating,
the difference came back exactly 0.0, the 1e-12 guard took over, and the
metric -- documented as bounded in [0, 1] -- hit 612, returning a
start_index ~41,000 samples off. Reproduced at 35 and 40 dB.

The numba path was never affected: it slides its window with float64
add/subtract instead of differencing prefix sums. This was worst for
backend="cupy", which deliberately skips numba and so had no correct
path at all; on x86 numba normally masks it.

The existing numpy-vs-numba cross-checks missed this because they
compare numba against a local COPY of the computation inside the test
file, which tracked the defect rather than catching it. That copy is now
kept in sync, and a new test drives the real dispatch on a
full-length frame with a quiet tail -- verified to fail before this fix
and pass after.

Pre-existing unrelated failures (hls/, mac_session, one ofdm_class FEC
test) are unchanged at 14; passing tests go 1394 -> 1411.

+++ spectracuda/sync/schmidl_cox.py
@@ -114,7 +114,42 @@ class SchmidlCoxSync(Block):
         # candidate d at once via prefix sums (no per-offset Python loop).
         # R(d) symmetrizes over BOTH halves' energy -- see module
         # docstring for why the textbook second-half-only R(d) is unsafe.
-        a = xp.conj(rx[:, :-L]) * rx[:, L:]
+        # DOUBLE precision for the prefix sums, not the input's
+        # complex64/float32. A prefix sum is a running total over the WHOLE
+        # buffer, and every windowed sum below is recovered by differencing
+        # two of those totals -- so the accumulator's ulp, not the window's
+        # own magnitude, sets the smallest window sum that survives.
+        #
+        # In float32 that breaks outright on any quiet stretch: on a real
+        # 44,672-sample frame the energy total reaches ~137, whose float32
+        # ulp is ~1.6e-5, while a sample in the low-energy region past the
+        # frame is ~3.5e-7 -- roughly 46,000x below half-ulp, so those
+        # samples stop accumulating ENTIRELY and the difference comes back
+        # as exactly 0.0 instead of ~4.5e-5. The `+ 1e-12` guard below then
+        # takes over and metric = |P|^2/1e-12 explodes to ~62 -- on a
+        # quantity the module docstring (correctly) documents as bounded in
+        # [0, 1] -- handing back a start_index tens of thousands of samples
+        # from the true peak. Measured at 35 and 40 dB SNR before this fix;
+        # covered by tests/test_sync_numba_acceleration.py::
+        # test_shipped_numpy_path_matches_numba_at_high_snr (which drives
+        # the real dispatch, unlike that file's older _numpy_reference
+        # cross-checks -- those compare numba against a local COPY of this
+        # computation, so they tracked the bug instead of catching it).
+        #
+        # This is the same "spurious spike at a low-energy boundary" the
+        # module docstring describes; symmetrizing R(d) narrowed when it
+        # fires but did not remove it, because the cause is the
+        # accumulator, not the formula. The `numba_process` path above was
+        # never affected: it slides its window with float64 add/subtract
+        # (Python complex/float accumulators) rather than differencing
+        # prefix sums, which is why the two paths disagreed. Doubling here
+        # makes the vectorized path match the kernel's precision -- it is
+        # NOT merely a tighter tolerance, it is the difference between a
+        # window sum existing and vanishing.
+        #
+        # Matters most for backend="cupy", which deliberately skips the
+        # numba path (see above) and so had no correct path at all.
+        a = (xp.conj(rx[:, :-L]) * rx[:, L:]).astype("complex128")
         n_batch = rx.shape[0]
 
         def _cumsum_with_leading_zero(x):
@@ -147,7 +182,7 @@ class SchmidlCoxSync(Block):
         #     metric correctness on real synthetic preambles, not just
         #     numerical closeness of an intermediate) still passes
         #     unchanged.
-        e = xp.abs(rx) ** 2
+        e = (xp.abs(rx) ** 2).astype("float64")  # float64: see `a` above
         e_cum = _cumsum_with_leading_zero(e)
         b1_cum = e_cum[:, : n_samples - L + 1]
         b2_cum = e_cum[:, L:] - e_cum[:, L : L + 1]
